Update setup.sh file
WSL does not install docker.

+++ etc/setup.sh
@@ -119,17 +119,19 @@ echo "****************"
 # docker
 echo "***** docker *****"
 if ! has "docker"; then
-  sudo apt install -y apt-transport-https ca-certificates curl gnupg-agent software-properties-common
-  curl -fsSL https://download.docker.com/linux/ubuntu/gpg | sudo apt-key add -
-  sudo add-apt-repository "deb [arch=amd64] https://download.docker.com/linux/ubuntu $(lsb_release -cs) stable"
-  sudo apt update
-  sudo apt install -y docker-ce docker-ce-cli containerd.io
-  sudo gpasswd -a ${USER} docker
-  cat /etc/group | grep docker
-  sudo chmod 666 /var/run/docker.sock
-  compose_version="$(curl -s https://api.github.com/repos/docker/compose/releases/latest | grep tag_name | sed -e 's/[^0-9\.]//g')"
-  sudo curl -L "https://github.com/docker/compose/releases/download/${compose_version}/docker-compose-$(uname -s)-$(uname -m)" -o /usr/local/bin/docker-compose
-  sudo chmod +x /usr/local/bin/docker-compose
+  if [ ! -f /proc/sys/fs/binfmt_misc/WSLInterop ]; then
+    sudo apt install -y apt-transport-https ca-certificates curl gnupg-agent software-properties-common
+    curl -fsSL https://download.docker.com/linux/ubuntu/gpg | sudo apt-key add -
+    sudo add-apt-repository "deb [arch=amd64] https://download.docker.com/linux/ubuntu $(lsb_release -cs) stable"
+    sudo apt update
+    sudo apt install -y docker-ce docker-ce-cli containerd.io
+    sudo gpasswd -a ${USER} docker
+    cat /etc/group | grep docker
+    sudo chmod 666 /var/run/docker.sock
+    compose_version="$(curl -s https://api.github.com/repos/docker/compose/releases/latest | grep tag_name | sed -e 's/[^0-9\.]//g')"
+    sudo curl -L "https://github.com/docker/compose/releases/download/${compose_version}/docker-compose-$(uname -s)-$(uname -m)" -o /usr/local/bin/docker-compose
+    sudo chmod +x /usr/local/bin/docker-compose
+  fi
 else
   docker --version
   docker-compose --version
